requirements.txt
boto3==1.4.4
django==1.11.4
django-ajax-selects==1.6.0
django-ckeditor==5.2.2
django-datatables-view==1.14.0
django-daterange-filter==1.3.0
django-environ==0.4.3
django-import-export==0.5.1
django-localflavor==1.5.1
django-object-actions==0.10.0
django-storages==1.5.2
django-stronghold==0.2.8
django-tabular-export==1.0.2
gunicorn==19.7.1
haversine==0.4.5
openpyxl==2.4.7
psycopg2==2.7.1
python-dateutil==2.6.0
requests==2.18.1
whitenoise==3.3.0  # used by heroku
xlsxwriter==0.9.6
